+++ src/Symfony/Component/Workflow/Tests/RegistryTest.php
@@ -58,8 +58,8 @@ class RegistryTest extends TestCase
         $workflow = $this->registry->get(Subject1::class);
         $this->assertInstanceOf(Workflow::class, $workflow);
         $this->assertSame('workflow1', $workflow->getName());
-
     }
+
     public function testGetWithMultipleMatch()
     {
         $this->expectException('Symfony\Component\Workflow\Exception\InvalidArgumentException');
@@ -110,7 +110,7 @@ class RegistryTest extends TestCase
         $strategy = $this->getMockBuilder(WorkflowSupportStrategyInterface::class)->getMock();
         $strategy->expects($this->any())->method('supports')
             ->willReturnCallback(function ($workflow, $subject) use ($supportedClassName) {
-                if (is_object($subject)) {
+                if (\is_object($subject)) {
                     return $subject instanceof $supportedClassName;
                 } else {
                     return $subject === $supportedClassName;